Move all setup-related tests into the same file
